Dynamic Trackbars now working too
**IMPORTANT**
Player is now a member of PlayerGUI instead of
PlayerGui being a member of Player!

+++ SoundBoard/SoundBoard/WaveForm.cpp
@@ -48,13 +48,13 @@ namespace SoundBoard
 		}
 		
 		//deletes the raw-file that was previously created
-		/*if(File::Exists(Environment::CurrentDirectory + "\\sox\\" + temp + ".raw"))
+		if(File::Exists(Environment::CurrentDirectory + "\\sox\\" + temp + ".raw"))
 		{
 
 			File::Delete(Environment::CurrentDirectory + "\\sox\\" + temp + ".raw");
 
 		}
-		*/
+		
 		return bmp;
 	}
